+++ src/Telegram.Bot/Types/ReplyMarkups/Buttons/IKeyboardButton.cs
@@ -1,0 +1,13 @@
+﻿namespace Telegram.Bot.Types.ReplyMarkups.Buttons
+{
+    /// <summary>
+    /// This interface represents one button of the reply keyboard (regular or inline)
+    /// </summary>
+    public interface IKeyboardButton
+    {
+        /// <summary>
+        /// Label text on the button
+        /// </summary>
+        string Text { get; set; }
+    }
+}