+++ test/Telegram.Bot.Tests.Integ/Framework/TestsFixture.cs
@@ -4,8 +4,8 @@ using System.Threading;
 using System.Threading.Tasks;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.Enums;
-using Telegram.Bot.Types.InlineKeyboardButtons;
 using Telegram.Bot.Types.ReplyMarkups;
+using Telegram.Bot.Types.ReplyMarkups.Buttons;
 using Xunit.Sdk;
 
 namespace Telegram.Bot.Tests.Integ.Framework