+++ test/Telegram.Bot.Tests.Integ/Location/InlineMessageLiveLocationTests .cs
@@ -2,9 +2,9 @@ using System;
 using System.Threading.Tasks;
 using Telegram.Bot.Tests.Integ.Framework;
 using Telegram.Bot.Types;
-using Telegram.Bot.Types.InlineKeyboardButtons;
 using Telegram.Bot.Types.InlineQueryResults;
 using Telegram.Bot.Types.ReplyMarkups;
+using Telegram.Bot.Types.ReplyMarkups.Buttons;
 using Xunit;
 
 // ReSharper disable once CheckNamespace